refactor: remove redundant type conversion

+++ controller.go
@@ -357,7 +357,7 @@ func NewCustomController(
 ) *Controller {
 	ctrlID := registerCustomControllerCallbacks(ctrl)
 	handle := maa.MaaCustomControllerCreate(
-		uintptr(unsafe.Pointer(customControllerCallbacksHandle)),
+		unsafe.Pointer(customControllerCallbacksHandle),
 		// Here, we are simply passing the uint64 value as a pointer
 		// and will not actually dereference this pointer.
 		uintptr(ctrlID),
